feat: enhance app usage chart and add force reload
This commit introduces several improvements to the app usage chart and adds a "Force Reload" option to the widget settings.

The following changes were made:
- Updated `AppUsageChart` to display a legend for CPU time and CPU foreground time.
- Increased the height of the `AppUsageChart` in `AppUsageList`.
- Centered the percentage text in `BatteryItem` when the device name is empty.
- Added new string resources for CPU time, CPU foreground, and usage time in English, Japanese, Vietnamese, and French.
- Added a "Force Reload" option to the dropdown menu in `DropdownMenu.kt` to allow users to manually refresh widget data.
- Modified `timeMillisToHours` to return a `Float` for more precise calculations.
- Added `@RequiresApi(Build.VERSION_CODES.Q)` annotations to `AppUsageScreen` and related composables as app usage statistics are only available from Android Q.
- Updated `sendMail` in `ExternalControllerExt.kt` to use `toUri()` extension.

+++ app/src/main/java/io/github/ifa/glancewidget/utils/CommonExt.kt
@@ -178,6 +178,6 @@ suspend fun Context.updateBatteryWidget() {
     }
 }
 
-fun Long.timeMillisToHours(): Long {
-    return this / (1000 * 60 * 60)
+fun Long.timeMillisToHours(): Float {
+    return (this.toDouble() / (1000 * 60 * 60)).toFloat()
 }
